E0-I12 убрал замечания Гита (изменил имена пакетов)
