+++ app/src/main/java/ru/practicum/android/diploma/presentation/country/SelectCountryViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.places_of_work
+package ru.practicum.android.diploma.presentation.country
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class PlacesOfWorkViewModel : ViewModel() {
+class SelectCountryViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
