+++ app/src/main/java/ru/practicum/android/diploma/presentation/filter/FilteringSettingsViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.region_selection
+package ru.practicum.android.diploma.presentation.filter
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class RegionSelectionViewModel : ViewModel() {
+class FilteringSettingsViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
