+++ app/src/main/java/ru/practicum/android/diploma/presentation/industry/IndustrySelectionViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.job_vacancy
+package ru.practicum.android.diploma.presentation.industry
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class JobVacancyViewModel : ViewModel() {
+class IndustrySelectionViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
