+++ app/src/main/java/ru/practicum/android/diploma/presentation/place/PlacesOfWorkViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.filtering_settings
+package ru.practicum.android.diploma.presentation.place
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class FilteringSettingsViewModel : ViewModel() {
+class PlacesOfWorkViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
