+++ app/src/main/java/ru/practicum/android/diploma/presentation/region/RegionSelectionViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.industry_selection
+package ru.practicum.android.diploma.presentation.region
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class IndustrySelectionViewModel : ViewModel() {
+class RegionSelectionViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
