+++ app/src/main/java/ru/practicum/android/diploma/presentation/vacancy/JobVacancyViewModel.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.presentation.select_country
+package ru.practicum.android.diploma.presentation.vacancy
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SelectCountryViewModel : ViewModel() {
+class JobVacancyViewModel : ViewModel() {
 
     private val _liveData = MutableLiveData<String>()
     val liveData: LiveData<String> = _liveData
