+++ app/src/main/java/ru/practicum/android/diploma/ui/team/TeamPageFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.team_page
+package ru.practicum.android.diploma.ui.team
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.databinding.FragmentTeamPageBinding
 
-class TeamPageFragment() : Fragment() {
+class TeamPageFragment : Fragment() {
 
     private var binding: FragmentTeamPageBinding? = null
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
