+++ app/src/main/res/layout/fragment_filtering_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.filtering_settings.FilteringSettingsFragment">
+    tools:context=".ui.filter.FilteringSettingsFragment">
 
     <TextView
         android:layout_width="wrap_content"
