+++ app/src/main/res/layout/fragment_industry_selection.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.industry_selection.IndustrySelectionFragment">
+    tools:context=".ui.industry.IndustrySelectionFragment">
 
     <TextView
         android:layout_width="wrap_content"
