+++ app/src/main/res/layout/fragment_job_vacancy.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.job_vacancy.JobVacancyFragment">
+    tools:context=".ui.vacancy.JobVacancyFragment">
 
     <TextView
         android:layout_width="wrap_content"
