+++ app/src/main/res/layout/fragment_places_of_work.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.places_of_work.PlacesOfWorkFragment">
+    tools:context=".ui.place.PlacesOfWorkFragment">
 
     <TextView
         android:layout_width="wrap_content"
