+++ app/src/main/res/layout/fragment_region_selection.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.region_selection.RegionSelectionFragment">
+    tools:context=".ui.region.RegionSelectionFragment">
 
     <TextView
         android:layout_width="wrap_content"
