+++ app/src/main/res/layout/fragment_select_country.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.select_country.SelectCountryFragment">
+    tools:context=".ui.country.SelectCountryFragment">
 
     <TextView
         android:layout_width="wrap_content"
