+++ app/src/main/res/layout/fragment_team_page.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.team_page.TeamPageFragment">
+    tools:context=".ui.team.TeamPageFragment">
 
     <TextView
         android:layout_width="wrap_content"
